+++ RayLibShowcase11/src/Tree.cpp
@@ -28,5 +28,5 @@ void Tree::Shape(cp::Space *mSpace) {
 }
 
 void Tree::draw() {
-    DrawTexturePro(texture2D, {0, 0, texture2D.width, texture2D.height}, {position.x, position.y, scale.x, scale.y}, {0, 0}, 0, WHITE);
+    DrawTexturePro(texture2D, {0, 0, (float)texture2D.width, (float)texture2D.height}, {position.x, position.y, scale.x, scale.y}, {0, 0}, 0, WHITE);
 }
